refactor(turns): decompose row_store.rs into write_behind / load / commit submodules
`row_store.rs` (the row-store module root) was ~1,597 lines, over the 1,500
threshold and carrying two `arch-exempt: large_file` markers. Split the big
`impl FilesystemTurnStateRowStore` into three cohesive child modules:

- row_store.rs (527) — struct/enum defs (incl. RowCollection), constructors +
  builders, observability accessors, snapshot-cache read helpers, runner-lease
  glue, span helpers, mod/use decls.
- row_store/commit.rs (490) — the two apply engines (`apply`,
  `apply_with_targeted_delta`), overlay acquisition, and the run-state
  transition wrappers.
- row_store/load.rs (411) — durable rehydration, legacy-blob migration, and the
  materialized-row / durable readers.
- row_store/write_behind.rs (247) — the write-behind window: reserve / enqueue /
  track / commit_pending / await / drain / degradation guards.

Move-only: struct fields stay in row_store.rs so submodules keep field access;
only cross-module methods were widened from private to `pub(super)` (still
crate-internal). No public API change; every file is now under 900 lines and
both `arch-exempt: large_file` markers are dropped. ironclaw_turns suite
(15 binaries) + clippy --all-targets --all-features + fmt --check green.
